requirements.txt
grpcio
grpcio-tools
fastapi
uvicorn